refactor: extract shared app layer (UART/clock/init) so the demos stay minimal
- new inc/app.h + src/app.c: lossless UART TX ring buffer, app_init()
  bundles clock+UART+LED setup, safe uart_write_int (full int32 range),
  ds18b20_busy strong override
- demo.c/demo2.c: only demo logic remains (callback + main), UART is now
  lossless in both, banners flushed before main loop
- Makefile: add src/app.c, per-app UART_TX_BUF_SIZE (demo 128, demo2 256)
